gitlab-ci, build/windows: Make Build script compatible with Local-Native build
The local builds now can use the same script of CI for sake of time.
The compatibility layer is very simple and makes the script more
used/tested, therefore more reliable.

+++ build/windows/gitlab-ci/2_build-gimp-msys2.sh
@@ -1,49 +1,62 @@
 #!/bin/bash
-# $MINGW_PACKAGE_PREFIX is defined by MSYS2.
+# $MSYSTEM_CARCH, $MSYSTEM_PREFIX and $MINGW_PACKAGE_PREFIX are defined by MSYS2.
 # https://github.com/msys2/MSYS2-packages/blob/master/filesystem/msystem
 
 set -e
 
-if [[ "$MSYSTEM" == "MINGW32" ]]; then
-    export ARTIFACTS_SUFFIX="-w32"
-    export MSYS2_ARCH_FOLDER="mingw32"
-    export MSYS2_PREFIX="/c/msys64/mingw32"
-elif [[ "$MSYSTEM" == "MINGW64" ]]; then
-    export ARTIFACTS_SUFFIX="-w64"
-    export MSYS2_ARCH_FOLDER="mingw64"
-    export MSYS2_PREFIX="/c/msys64/mingw64/"
-else # [[ "$MSYSTEM" == "CLANGARM64" ]];
-    export ARTIFACTS_SUFFIX="-arm64"
-    export MSYS2_ARCH_FOLDER="clangarm64"
-    export MSYS2_PREFIX="/c/msys64/clangarm64/"
+if [[ "$MSYSTEM_CARCH" == "aarch64" ]]; then
+    export ARTIFACTS_SUFFIX="-a64"
+    export MSYS2_PREFIX="c:/msys64${MSYSTEM_PREFIX}"
+elif [[ "$MSYSTEM_CARCH" == "x86_64" ]]; then
+    export ARTIFACTS_SUFFIX="-x64"
+    export MSYS2_PREFIX="c:/msys64${MSYSTEM_PREFIX}"
+else # [[ "$MSYSTEM_CARCH" == "i686" ]];
+    export ARTIFACTS_SUFFIX="-x86"
+    export MSYS2_PREFIX="c:/msys64${MSYSTEM_PREFIX}"
 fi
 
-export PATH="${MSYS2_PREFIX}/bin:$PATH"
+if [[ "$BUILD_TYPE" == "CI_NATIVE" ]]; then
+  # XXX We've got a weird error when the prefix is in the current dir.
+  # Until we figure it out, this trick seems to work, even though it's
+  # completely ridiculous.
+  rm -fr ~/_install${ARTIFACTS_SUFFIX}
+  mv "_install${ARTIFACTS_SUFFIX}" ~
+else
+  # Make the script callable from every directory
+  if [[ "$0" != "build/windows/gitlab-ci/2_build-gimp-msys2.sh" ]]; then
+    GIMP_EXTDIR="$0"
+    GIMP_EXTDIR=$(sed 's|build/windows/gitlab-ci/2_build-gimp-msys2.sh||g' <<< $GIMP_EXTDIR)
+    cd $GIMP_EXTDIR
+  else
+    GIMP_GITDIR="$(pwd)"
+    GIMP_GITDIR=$(sed 's|build/windows/gitlab-ci||g' <<< $GIMP_GITDIR)
+    GIMP_GITDIR=$(sed 's|build/windows||g' <<< $GIMP_GITDIR)
+    GIMP_GITDIR=$(sed 's|build||g' <<< $GIMP_GITDIR)
+    cd $GIMP_GITDIR
+  fi
+fi
 
-# Update everything
-pacman --noconfirm -Suy
 
 # Install the required (pre-built) packages for GIMP
+pacman --noconfirm -Suy
+
 export DEPS_PATH="build/windows/gitlab-ci/all-deps-uni.txt"
 sed -i "s/DEPS_ARCH_/${MINGW_PACKAGE_PREFIX}-/g" $DEPS_PATH
 export GIMP_DEPS=`cat $DEPS_PATH`
 pacman --noconfirm -S --needed base-devel $GIMP_DEPS
 
-# XXX We've got a weird error when the prefix is in the current dir.
-# Until we figure it out, this trick seems to work, even though it's
-# completely ridiculous.
-rm -fr ~/_install${ARTIFACTS_SUFFIX}
-mv "_install${ARTIFACTS_SUFFIX}" ~
 
+# Build GIMP
 export GIMP_PREFIX="`realpath ~/_install`${ARTIFACTS_SUFFIX}"
 export PATH="$GIMP_PREFIX/bin:$PATH"
 export PKG_CONFIG_PATH="${GIMP_PREFIX}/lib/pkgconfig:$PKG_CONFIG_PATH"
 export PKG_CONFIG_PATH="${GIMP_PREFIX}/share/pkgconfig:$PKG_CONFIG_PATH"
 export LD_LIBRARY_PATH="${GIMP_PREFIX}/lib:${LD_LIBRARY_PATH}"
-export ACLOCAL_FLAGS="-I/c/msys64/${MSYS2_ARCH_FOLDER}/share/aclocal"
-export XDG_DATA_DIRS="${GIMP_PREFIX}/share:/${MSYS2_ARCH_FOLDER}/share/"
+export XDG_DATA_DIRS="${GIMP_PREFIX}/share:${MSYSTEM_PREFIX}/share/"
 
-mkdir -p _ccache
+if [ ! -d "_ccache" ]; then
+  mkdir -p _ccache
+fi
 export CCACHE_BASEDIR="$(pwd)"
 export CCACHE_DIR="${CCACHE_BASEDIR}/_ccache"
 
@@ -55,23 +68,30 @@ export CCACHE_DIR="${CCACHE_BASEDIR}/_ccache"
 #ccache --zero-stats
 #ccache --show-stats
 
-mkdir "_build${ARTIFACTS_SUFFIX}"
-cd "_build${ARTIFACTS_SUFFIX}"
-# We disable javascript as we are not able for the time being to add a
-# javascript interpreter with GObject Introspection (GJS/spidermonkey
-# and Seed/Webkit are the 2 contenders so far, but they are not
-# available on MSYS2 and we are told it's very hard to build them).
-# TODO: re-enable javascript plug-ins when we can figure this out.
-meson .. -Dprefix="${GIMP_PREFIX}"           \
-         -Ddirectx-sdk-dir="${MSYS2_PREFIX}" \
-         -Djavascript=disabled               \
-         -Dbuild-id=org.gimp.GIMP_official   \
-         -Dgi-docgen=disabled
-ninja
-ninja install
-cd ..
+if [ ! -f "_build${ARTIFACTS_SUFFIX}/build.ninja" ]; then
+  mkdir -p "_build${ARTIFACTS_SUFFIX}" && cd "_build${ARTIFACTS_SUFFIX}"
+  # We disable javascript as we are not able for the time being to add a
+  # javascript interpreter with GObject Introspection (GJS/spidermonkey
+  # and Seed/Webkit are the 2 contenders so far, but they are not
+  # available on MSYS2 and we are told it's very hard to build them).
+  # TODO: re-enable javascript plug-ins when we can figure this out.
+  meson setup .. -Dprefix="${GIMP_PREFIX}"           \
+                 -Dgi-docgen=disabled                \
+                 -Djavascript=disabled               \
+                 -Ddirectx-sdk-dir="${MSYS2_PREFIX}" \
+                 -Dbuild-id=org.gimp.GIMP_official
+  ninja && ninja install
+else
+  cd "_build${ARTIFACTS_SUFFIX}"
+  ninja && ninja install
+fi
 
-#ccache --show-stats
 
-# XXX Moving back the prefix to be used as artifacts.
-mv "${GIMP_PREFIX}" .
+if [[ "$BUILD_TYPE" == "CI_NATIVE" ]]; then
+  cd ..
+
+  #ccache --show-stats
+
+  # XXX Moving back the prefix to be used as artifacts.
+  mv "${GIMP_PREFIX}" .
+fi
